+++ docs/changelog.md
@@ -8,6 +8,10 @@ will continue to work with version 0.6.1, but may break with version 0.7.0.
 
 ## Unreleased
 
+No changes yet!
+
+## v0.8.1
+
 - ADD: {meth}`cernml.coi.Config.extend()` to make configuration more composable.
 - ADD: {class}`cernml.coi.ConfigValues` as a convenience alias for {class}`types.SimpleNamespace`.
 - ADD: {func}`~cernml.coi.checkers.check_configurable()` for all implementors of the {class}`~cernml.coi.Configurable` interface.